Get build scripts for reactjs working and move gulp build scripts to _BuildScripts directory
