+++ generators/spa/templates/Edit.html
@@ -1,10 +1,4 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <title></title>
-</head>
-<body>
-
-</body>
-</html>
+﻿[JavaScript:{ path: "~/DesktopModules/<%= namespace %>/<%= moduleName %>/Resources/js/edit.js"}]
+[CSS:{ path: "~/DesktopModules/<%= namespace %>/<%= moduleName %>/Resources/bootstrap/css/bootstrap.min.css"}]
+[CSS:{ path: "~/DesktopModules/<%= namespace %>/<%= moduleName %>/Resources/module.css"}]
+<div class="<%= namespace.toLowerCase() %><%= moduleName %>Edit" data-moduleid="[ModuleContext:ModuleId]"></div>
