+++ generators/spa/templates/_Project.csproj
@@ -105,8 +105,7 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="package.json" />
-    <Content Include=".babelrc" />
-    <Content Include="manifest.dnn" />
+    <Content Include="<%= moduleName %>.dnn" />
     <Content Include="License.md" />
     <None Include="packages.config" />
     <Content Include="Providers\DataProviders\SqlDataProvider\01.00.00.SqlDataProvider" />
