+++ generators/spa/templates/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="DotNetNuke.Core" version="9.1.1.129" targetFramework="net461" />
   <package id="DotNetNuke.Web" version="9.1.1.129" targetFramework="net461" />
-  <package id="DotNetNuke.Web.Mvc" version="9.1.1.129" targetFramework="net461" />
+  <package id="DotNetNuke.WebApi" version="9.1.1.129" targetFramework="net461" />
   <package id="Microsoft.AspNet.Mvc" version="5.1.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.Razor" version="3.1.2" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebPages" version="3.1.2" targetFramework="net461" />
